Refactor AssetDaemonScenarioState 2
