+++ python_modules/dagster/dagster_tests/definitions_tests/auto_materialize_tests/updated_scenarios/latest_materialization_run_tag_scenarios.py
@@ -1,8 +1,8 @@
 from dagster import AutoMaterializeAssetPartitionsFilter, AutoMaterializePolicy, AutoMaterializeRule
 
-from ..asset_daemon_scenario import AssetDaemonScenario
 from ..base_scenario import run_request
-from .asset_daemon_scenario_states import (
+from ..utils.asset_daemon_scenario import AssetDaemonScenario
+from ..utils.asset_scenario_states import (
     one_asset_depends_on_two,
     three_assets_in_sequence,
     two_assets_in_sequence,
