+++ python_modules/dagster/dagster_tests/definitions_tests/auto_materialize_tests/utils/__init__.py
@@ -1,0 +1,18 @@
+import datetime
+
+from dagster import MultiPartitionKey
+
+
+def day_partition_key(time: datetime.datetime, delta: int = 0) -> str:
+    """Returns the partition key of a day partition delta days from the initial time."""
+    return (time + datetime.timedelta(days=delta - 1)).strftime("%Y-%m-%d")
+
+
+def hour_partition_key(time: datetime.datetime, delta: int = 0) -> str:
+    """Returns the partition key of a day partition delta days from the initial time."""
+    return (time + datetime.timedelta(hours=delta - 1)).strftime("%Y-%m-%d-%H:00")
+
+
+def multi_partition_key(**kwargs) -> MultiPartitionKey:
+    """Returns a MultiPartitionKey based off of the given kwargs."""
+    return MultiPartitionKey(kwargs)
